+++ src/components/modules/Bible.vue
@@ -1,8 +1,8 @@
 <template>
-  <q-card>
+  <q-card color="secondary">
     <div v-show="!data.editing || data.editing !== $firebase.auth.currentUser.uid">
       <q-card-title>
-        <q-icon v-show="!data.editing" class="float-right cursor-pointer" name="fas fa-ellipsis-v" color="primary" size="1rem">
+        <q-icon v-show="!data.editing" class="float-right cursor-pointer" name="fas fa-ellipsis-v" size="1rem">
           <q-popover>
             <q-list>
               <q-item link v-close-overlay @click.native="edit(id)">Edit</q-item>
@@ -11,7 +11,7 @@
           </q-popover>
         </q-icon>
         <span class="float-right" style="font-size: .8rem; vertical-align: top; line-height: 1rem;">{{ data.time }} minutes&nbsp;&nbsp;&nbsp;</span>
-        <q-icon name="fas fa-book" color="primary" size="2rem" />&nbsp;&nbsp;&nbsp;
+        <q-icon name="fas fa-book" size="2rem" />&nbsp;&nbsp;&nbsp;
         {{ data.bibleRef }} <span class="q-caption">{{ data.translation }}</span>
       </q-card-title>
       <q-card-main>
@@ -20,21 +20,22 @@
     </div>
     <div v-if="data.editing === $firebase.auth.currentUser.uid">
       <q-card-main>
-        <q-icon link class="float-right cursor-pointer" name="fas fa-times" color="primary" size="1rem" @click.native="close" />
+        <q-icon link class="float-right cursor-pointer" name="fas fa-times" size="1rem" @click.native="close" />
         <div class="row gutter-sm">
           <div class="col-12">
-            <q-input v-model="data.bibleRef" float-label="Bible Ref" @keyup.enter="preSave" />
+            <q-input v-model="data.bibleRef" float-label="Bible Ref" @keyup.enter="preSave" dark />
           </div>
           <div class="col-12">
             <q-select
               v-model="translation"
               float-label="Bible Translation"
               :options="translationOptions"
+              dark
             />
           </div>
           <div class="col-12">
             <q-btn color="primary" @click.native="preSave">Save</q-btn>
-            <q-btn color="negative" @click.native="remove(id)">Delete</q-btn>
+            <q-btn outline color="negative" @click.native="remove(id)">Delete</q-btn>
           </div>
         </div>
       </q-card-main>